src/app/pages/discussion/discussion-detail/discussion-detail.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute } from '@angular/router';
import { DiscussionService } from '../../../core';

@Component({
  selector: 'app-discussion-detail',
  templateUrl: './discussion-detail.component.html',
  styleUrls: ['./discussion-detail.component.scss']
})
export class DiscussionDetailComponent implements OnInit {
  discussion: any;
  id: string;

  constructor(
    private route: ActivatedRoute,
    private discussionService: DiscussionService
  ) {
    this.id = this.route.snapshot.paramMap.get('id');
  }

  ngOnInit() {
    this.discussion = this.discussionService.get(this.id);
  }

}
